Identity Role test and handle errors with middleware

+++ DigitalStore.Api/DigitalStore.Data/Domain/OrderDetail.cs
@@ -9,9 +9,9 @@ namespace DigitalStore.Data.Domain
 {
     public class OrderDetail : BaseEntity
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public Product Product { get; set; }
-        public int OrderId { get; set; }
+        public long OrderId { get; set; }
         public Order Order { get; set; }
         public int Quantity { get; set; }
         public decimal UnitPrice { get; set; }
